Corrigido entrypoint para não rodar se não existir o arquivo GemFile

+++ docker-entrypoint.sh
@@ -5,7 +5,9 @@
 set -e
 
 # Ensure all gems installed. Add binstubs to bin which has been added to PATH in Dockerfile.
-bundle check || bundle install --binstubs="$BUNDLE_BIN"
+if [[ -f Gemfile ]]; then
+  bundle check || bundle install --binstubs="$BUNDLE_BIN"
+fi
 
 if [ $# -eq 0 ]; then
 	exec /bin/bash
